Bandwidth and Native Mode
In the Yammer Native Mode Migration explanations and considerations is not mentioned in any Microsoft documentation that our customers don´t need have bandwidth concerns with this needed migration. Already had customers, concerns about this which was blocking them to perform the Native Mode migration. 

In this sense can we add a Note/Line in the documentation saying something like: "There are no bandwidth requirements for the Native mode migration. Essentially you are moving between Yammer separate site to Microsoft 365/SharePoint Online, nothing is downloaded. The only consideration to have in for SharePoint Online quota, in case Yammer was heavily being used”.